UPDATE: test refactor
