+++ __tests__/unit/store/inventory-item/mocks/default.js
@@ -1,0 +1,16 @@
+export default {
+  $store: {
+    state: {
+      activeProduct: {
+        product: null,
+      },
+      cart: {
+        order: null,
+      },
+    },
+    actions: {
+      'activeProduct/setActiveProduct': jest.fn(),
+    },
+    dispatch: jest.fn(),
+  },
+};
